blkbox-flask_api/requirements.txt
absl-py==0.12.0
alabaster==0.7.12
appdirs==1.4.4
apturl==0.5.2
arrow==1.0.3
asn1crypto==0.24.0
astroid==2.5.3
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.5
Babel==2.9.0
backcall==0.2.0
binaryornot==0.4.4
black==20.8b1
bleach==3.3.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.3.7
command-not-found==0.3
cookiecutter==1.7.2
cryptography==2.1.4
cupshelpers==1.0
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distro-info===0.18ubuntu0.18.04.1
docutils==0.16
entrypoints==0.3
flake8==3.8.4
Flask==1.1.2
gensim==3.8.0
httplib2==0.9.2
idna==2.5
imagesize==1.2.0
importlib-metadata==3.10.1
inflection==0.5.1
intervaltree==3.1.0
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==1.1.0
jedi==0.17.2
jeepney==0.6.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.0.1
jsonify==0.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keyring==23.0.1
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mediapipe==0.8.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
nltk==3.6.1
numpy==1.19.3
numpydoc==1.1.0
oauth==1.0.1
olefile==0.45.1
opencv-python==4.5.1.48
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.0
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
pluggy==0.13.1
poyo==0.5.0
prompt-toolkit==3.0.18
protobuf==3.15.6
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycups==1.9.73
pydocstyle==6.0.0
pyflakes==2.2.0
Pygments==2.8.1
pygobject==3.26.1
pylint==2.7.4
pyls-black==0.4.6
pyls-spyder==0.3.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.1
PyQtWebEngine==5.12.1
pyRFC3339==1.0
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-slugify==4.0.1
pytz==2018.3
pyxdg==0.27
PyYAML==3.12
pyzmq==22.0.3
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.2
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.4.4
reportlab==3.4.0
requests==2.25.1
requests-unixsocket==0.1.5
requests2==2.16.0
rope==0.18.0
scikit-learn==0.24.1
scipy==1.5.4
SecretStorage==3.3.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
smart-open==5.0.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==5.0.0
spyder-kernels==2.0.1
system-service==0.3
systemd-python==234
testpath==0.4.4
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.1.0
three-merge==0.1.1
tinycss==0.4
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.0.2
unattended-upgrades==0.1
urllib3==1.21.1
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
wurlitzer==2.1.0
xkit==0.0.0
yapf==0.31.0
zipp==3.4.1
zope.interface==4.3.2
